Added enemy spawning mechanics and allowed for rooms to be irregular
Added a method to DirtController that can access any particle by index in O(1) time and used it to randomly select a particle each frame and check if it is under an object on the furniture layer to spawn enemies. This can also be used to loop through every particle one per frame and check if it is under a wall or out of bounds, and if it is, destroy it. This could be a good way to prevent rooms from being softlocked.

The dirtcontroller now requires floorboard children objects to determine the size of the room, and uses the weights of their areas in order to evenly distribute particles between them.